latest bacnet stack changes from sourceforge

+++ ports/pic18f6720/bi.c
@@ -28,7 +28,6 @@
 #include <stdbool.h>
 #include <stdint.h>
 #include <stdio.h>
-#include <string.h>
 #include "bacdef.h"
 #include "bacdcode.h"
 #include "bacenum.h"
@@ -114,7 +113,7 @@ char *Binary_Input_Name(
     static char text_string[16] = "";   /* okay for single thread */
 
     if (object_instance < MAX_BINARY_INPUTS) {
-        sprintf(text_string, "BI-%lu",  (unsigned long) object_instance);
+        sprintf(text_string, "BI-%lu", object_instance);
         return text_string;
     }
